% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_w_noise.R
\name{add_w_noise}
\alias{add_w_noise}
\title{add_w_noise
Add two numbers together plus some noise}
\usage{
add_w_noise(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
A numeric vector.
}
\description{
add_w_noise
Add two numbers together plus some noise
}
