% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.protein_mods}
\alias{print.protein_mods}
\title{printR.R}
\usage{
\method{print}{protein_mods}(x, ...)
}
\arguments{
\item{x}{A data frame containing amino acid sequences}
}
\value{
A character sequence.  Character sequence is a visual representation
of amino acid sequence, along with modifications and their location
}
\description{
Print method for protein_mods objects
}
